requirements.txt
django
djangorestframework
django-filter
drf-yasg
markdown

